+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added a customizable Loupe (magnifier) feature to the comic reader, which can be toggled via keyboard or mouse shortcuts (#194)
 * Added multi-select functionality (Ctrl/Cmd+click, Shift+click) to the bookshelf grid, allowing users to apply bulk actions (set tags, add to collections, remove from bookshelf) to multiple books at once (#195)
 * Added a setting to toggle automatic scrolling for bookshelf titles to help improve performance (#196)
+* Added series grouping and drill-down navigation in the bookshelf view (#200)
 
 ### Changed